+++ packages/backend/src/queryBuilder.ts
@@ -141,7 +141,9 @@ export const getCustomDimensionSql = ({
     compiledMetricQuery: CompiledMetricQuery;
     fieldQuoteChar: string;
     userAttributes: UserAttributeValueMap | undefined;
-}): { ctes: string[]; joins: string[]; selects: string[] } | undefined => {
+}):
+    | { ctes: string[]; joins: string[]; tables: string[]; selects: string[] }
+    | undefined => {
     const { customDimensions } = compiledMetricQuery;
 
     if (customDimensions === undefined || customDimensions.length === 0)
@@ -159,7 +161,10 @@ export const getCustomDimensionSql = ({
         return ` ${getCteReference(customDimension)} AS (
             SELECT
                 MIN(${dimension.compiledSql}) AS min_id,
-                MAX(${dimension.compiledSql}) AS max_id
+                MAX(${dimension.compiledSql}) AS max_id,
+                MIN(${dimension.compiledSql}) + (MAX(${
+            dimension.compiledSql
+        }) - MIN(${dimension.compiledSql}) ) as ratio
             FROM ${baseTable} AS ${fieldQuoteChar}${
             customDimension.table
         }${fieldQuoteChar}
@@ -168,6 +173,9 @@ export const getCustomDimensionSql = ({
 
     const joins = customDimensions.map(getCteReference);
 
+    const tables = customDimensions.map(
+        (customDimension) => customDimension.table,
+    );
     const selects = customDimensions.reduce<string[]>(
         (acc, customDimension) => {
             const dimension = getDimensionFromId(
@@ -194,7 +202,7 @@ export const getCustomDimensionSql = ({
                     }
 
                     // TODO test this on other warehouses
-                    const ratio = `${cte}.min_id + (${cte}.max_id - ${cte}.min_id )`;
+                    const ratio = `${cte}.ratio`;
 
                     if (customDimension.binNumber <= 1) {
                         // Edge case, bin number with only one bucket does not need a CASE statement
@@ -234,7 +242,7 @@ export const getCustomDimensionSql = ({
         [],
     );
 
-    return { ctes, joins, selects };
+    return { ctes, joins, tables: [...new Set(tables)], selects };
 };
 
 export const buildQuery = ({
@@ -309,6 +317,7 @@ export const buildQuery = ({
             const dim = getDimensionFromId(field, explore);
             return [...acc, ...(dim.tablesReferences || [dim.table])];
         }, []),
+        ...(customDimensionSql?.tables || []),
         ...getFilterRulesFromGroup(filters.dimensions).reduce<string[]>(
             (acc, filterRule) => {
                 const dim = getDimensionFromId(
@@ -357,7 +366,6 @@ export const buildQuery = ({
         );
         return [...allNewReferences, ...getJoinedTables(allNewReferences)];
     };
-
     const joinedTables = new Set([
         ...selectedTables,
         ...getJoinedTables([...selectedTables]),