+++ packages/frontend/src/components/Explorer/index.tsx
@@ -2,6 +2,8 @@ import { Stack } from '@mantine/core';
 import { FC, memo } from 'react';
 import { useParams } from 'react-router-dom';
 
+import { ProjectType } from '@lightdash/common';
+import { useProjects } from '../../hooks/useProjects';
 import { useExplorerContext } from '../../providers/ExplorerProvider';
 import DrillDownModal from '../MetricQueryData/DrillDownModal';
 import MetricQueryDataProvider from '../MetricQueryData/MetricQueryDataProvider';
@@ -24,6 +26,13 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
         );
         const { projectUuid } = useParams<{ projectUuid: string }>();
 
+        const { data: projects } = useProjects({ refetchOnMount: false });
+        const isProjectPreview = !!projects?.find(
+            (project) =>
+                project.projectUuid === projectUuid &&
+                project.type === ProjectType.PREVIEW,
+        );
+
         return (
             <MetricQueryDataProvider
                 metricQuery={unsavedChartVersionMetricQuery}
@@ -34,7 +43,10 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
 
                     <FiltersCard />
 
-                    <VisualizationCard projectUuid={projectUuid} />
+                    <VisualizationCard
+                        projectUuid={projectUuid}
+                        isProjectPreview={isProjectPreview}
+                    />
 
                     <ResultsCard />